+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unrelased]
 
+### Changed
+- Added log error when Microservice cannot start.
+
 ## [1.2.0]
 ### Added
 - Support for GUID based assembly references.